health-backend-java/health-backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
pydantic
python-multipart
python-dotenv
google-generativeai
